requirements.txt
flask>=2.2
flask-cors>=3.0
psycopg2-binary>=2.9
gunicorn>=20.1
pytest>=7.0
